fixed documentation and moved pubsub abc.py
