+++ docs/libp2p.network.connection.rst
@@ -12,26 +12,10 @@ libp2p.network.connection.exceptions module
    :undoc-members:
    :show-inheritance:
 
-libp2p.network.connection.net\_connection\_interface module
------------------------------------------------------------
-
-.. automodule:: libp2p.network.connection.net_connection_interface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 libp2p.network.connection.raw\_connection module
 ------------------------------------------------
 
 .. automodule:: libp2p.network.connection.raw_connection
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-libp2p.network.connection.raw\_connection\_interface module
------------------------------------------------------------
-
-.. automodule:: libp2p.network.connection.raw_connection_interface
    :members:
    :undoc-members:
    :show-inheritance:
