--- docs/libp2p.routing.rst
@@ -1,21 +1,0 @@
-libp2p.routing package
-======================
-
-Submodules
-----------
-
-libp2p.routing.interfaces module
---------------------------------
-
-.. automodule:: libp2p.routing.interfaces
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: libp2p.routing
-   :members:
-   :undoc-members:
-   :show-inheritance:
