+++ docs/libp2p.transport.rst
@@ -20,30 +20,6 @@ libp2p.transport.exceptions module
    :undoc-members:
    :show-inheritance:
 
-libp2p.transport.listener\_interface module
--------------------------------------------
-
-.. automodule:: libp2p.transport.listener_interface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-libp2p.transport.transport\_interface module
---------------------------------------------
-
-.. automodule:: libp2p.transport.transport_interface
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-libp2p.transport.typing module
-------------------------------
-
-.. automodule:: libp2p.transport.typing
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 libp2p.transport.upgrader module
 --------------------------------
